requirements.txt
antiorm==1.2.1
arrow==1.3.0
asgiref==3.8.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
db==0.1.1
db-sqlite3==0.0.1
deepdiff==8.0.1
dipy==1.9.0
Django==5.1.3
fonttools==4.54.1
h5py==3.12.1
image==1.5.33
kiwisolver==1.4.7
matplotlib==3.9.2
mysql-connector-python==9.1.0
nibabel==5.3.2
numpy==2.1.2
opencv-python==4.6.0.66
orderly-set==5.2.2
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-time==0.3.0
scipy==1.14.1
setuptools==75.5.0
setuptools-scm==8.1.0
six==1.16.0
sqlparse==0.5.2
TIME-python==0.0.17
tqdm==4.67.0
trx-python==0.3
types-python-dateutil==2.9.0.20241003
tzdata==2024.2
reportlab==4.3.1